+++ src/pcg/Noise.cpp
@@ -1,11 +1,12 @@
 // This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
+#include <span>
+#include <SDL2/SDL_render.h>
+
 #include "pcg/Noise.hpp"
 #include "pcg/Random.hpp"
 #include "utils/Logger.hpp"
-
-#include <SDL2/SDL_render.h>
 
 SDL_Texture *tnt::pcg::staticNoise(SDL_Renderer *ren, int w, int h)
 {
@@ -16,9 +17,10 @@ SDL_Texture *tnt::pcg::staticNoise(SDL_Renderer *ren, int w, int h)
     SDL_PixelFormat pixelFmt;
     pixelFmt.format = format;
 
-    Uint32 *pixels = new Uint32[w * h * 4];
+    std::vector<Uint32> pixels;
+    pixels.reserve(w * h * 4);
     int pitch{0};
-    if (SDL_LockTexture(ret, nullptr, (void **)&pixels, &pitch))
+    if (SDL_LockTexture(ret, nullptr, (void **)&pixels, &pitch) < 0)
     {
         logger::debug("staticNoise couldn't lock texture!!\nError: {}", SDL_GetError());
         return nullptr;
@@ -27,17 +29,15 @@ SDL_Texture *tnt::pcg::staticNoise(SDL_Renderer *ren, int w, int h)
     for (int i{0}; i < h; ++i)
         for (int j{0}; j < w; ++j)
         {
-            uint8_t r{static_cast<Uint8>(randomInt(0, 255))};
-            uint8_t g{static_cast<Uint8>(randomInt(0, 255))};
-            uint8_t b{static_cast<Uint8>(randomInt(0, 255))};
+            const uint8_t r{static_cast<Uint8>(randomInt(0, 255))};
+            const uint8_t g{static_cast<Uint8>(randomInt(0, 255))};
+            const uint8_t b{static_cast<Uint8>(randomInt(0, 255))};
 
             const Uint32 color{SDL_MapRGB(&pixelFmt, r, g, b)};
-
-            Uint32 pos{i * (pitch / sizeof(unsigned)) + j};
+            const Uint32 pos{i * (pitch / sizeof(unsigned)) + j};
             pixels[pos] = color;
         }
 
-    delete[] pixels;
     SDL_UnlockTexture(ret);
     return ret;
 }